[ADD] FirebaseModel 분리 (#50)
