+++ Hatnya/Network/Models/FirebaseModel/FirebaseStudyInfo.swift
@@ -1,0 +1,15 @@
+//
+//  FirebaseStudyInfo.swift
+//  Hatnya
+//
+//  Created by Mingwan Choi on 2022/07/31.
+//
+
+import Foundation
+
+struct FirebaseStudyInfo: Codable {
+    let name: String
+    let code: String
+    let createdAt: Date
+    let description: String
+}
