+++ Hatnya/Network/Models/FirebaseModel/HomeworkTask.swift
@@ -1,0 +1,13 @@
+//
+//  HomeworkTask.swift
+//  Hatnya
+//
+//  Created by Mingwan Choi on 2022/07/31.
+//
+
+import Foundation
+
+struct HomeworkTask: Codable {
+    var cycle: Int
+    var list: [Homework]
+}
